feat(TS/components):
- Use TS in Navigation component.

meta: Level-3
