+++ package.json
@@ -21,6 +21,7 @@
     "@types/node": "^18.13.0",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
+    "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "5.1.26",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "prop-types": "^15.8.1",
@@ -82,6 +83,10 @@
     "webpack": "^5.75.0"
   },
   "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ],
     "overrides": [
       {
         "files": [
