+++ tsconfig.json
@@ -8,6 +8,7 @@
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
     "outDir": "./build",
+    "suppressImplicitAnyIndexErrors": true
   },
   "include": [
     "src/**/*.ts",
